+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 CHANGELOG
 =========
 
-master
-------
+1.2.1
+-----
 
 Bug fixes: